requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
pillow
numpy
scikit-learn
